core: handle SIGUSR2 as a signal to reload the config
This PR adds the infrastructure to handle POSIX signals. The only signal
currently handled is SIGUSR2, which will reload the config.

SIGPIPE is set up to be ignored in the global state and should
not be changed as macOS relies on SIGPIPE being ignored for proper
operation.

Note that any signal not configured by Ghostty will generally result in
termination of the process.

Fixes #7747