linux: switch systemd user service to type notify-reload
This enables `systemd` to send SIGUSR2 to Ghostty to signal it to reload
the configuration. This is much easier and reliable than using a script
to search process listings for Ghostty's main PID to send the signal to.
The command to do so is:

`systemctl reload --user com.mitchellh.ghostty.service`