core/vt: initialize OSC command to a known value
Fixes: #8007

If we don't initialize the command value when we create an OSC parser,
there may be situations where the parser gets reset and it appears that
the command contains a `kitty_color_protocol` value (due to the value
Zig uses to initialize `undefined` values with). The program will then
crash when the ArrayList inside the `kitty_color_protocol` is attempted
to be deinitialized since the ArrayList was never properly initialized.

By always initializing the command to a known value, we avoid this
situation.